requirements.txt
streamlit
python-dotenv   ##(To store the API key securely instead of in os.environ)
langchain
langchain-openai
langchain-groq
langchain-google-genai
langchain-community